requirements.txt
pandas
matplotlib
wordcloud
textblob
pillow
streamlit
